+++ terraform/environments/production/main.tf
@@ -24,3 +24,10 @@ module "security_groups" {
 
   tags = local.common_tags
 }
+
+module "iam" {
+  source = "../../modules/iam"
+
+  project_name = var.project_name
+  tags         = local.common_tags
+}